Added an April Tag Motion Camera to the collection disk object
This is a camera that will be used to coordinate with rovers and help
them determine their position and could be used to help determine the
position of objects in a given rectangular space.

The next step of this feature will be integrating the camera into
the rover interface so that users can see camera data.